fix: segfault when getcwd is called with null buf
ruby and I'm guessing other apps pass a null buffer to getcwd, expecting
to use the return value rather than having data written into the
destination buffer. this obviously causes an illegal access. this patch
fixes it by having fakedir's getcwd check if the destination buffer is
valid and allocating a new buffer with strdup if it isn't.